Las opciones de administrador unicamente son mostradas a estos

+++ PHP/articulos.php
@@ -123,9 +123,14 @@
 					<td><input type='text' name='campo_codigo' value='".$articulo['codigo']."' readonly></td>
 					<td><input type='text' name='campo_descripcion' value='".$articulo['descripcion']."' readonly></td>
 					<td><input type='text' name='campo_observaciones' value='".$articulo['observaciones']."' readonly></td>
-					<td>
+					<td>";
+			if ($_SESSION === "administrador") {
+				echo "
 						<button onclick='eliminar_articulo(this)' type='button' data-toggle='tooltip' data-placement='top' title='Eliminar artículo'><i class='fas fa-trash'></i></button>
 						<button onclick='modificar_articulo(this)' type='button' data-toggle='tooltip' data-placement='top' title='Modificar artículo'><i class='fas fa-pen'></i></button>
+				";
+			}
+			echo "
 					</td>
 				</tr>
 			";
